Cleanup API errors
