+++ lib/anthropic/api/messages.rb
@@ -8,7 +8,7 @@ module Anthropic
       def create(**params, &)
         streaming = params[:stream]
         if streaming && beta_loaded?('tools-2024-04-04')
-          raise Anthropic::Errors::UnsupportedBetaUseError, 'Tool use is not yet supported in streaming mode'
+          raise Anthropic::Api::UnsupportedBetaUseError, 'Tool use is not yet supported in streaming mode'
         end
 
         validate!(params)
